Don't randomly remove the reviewed patches
Only remove them if they are a temporary file created for a command.
This class should be probably splitted in two because the command case
is clearly a specific case of the file, at least in this implementation.

+++ plugins/patchreview/localpatchsource.cpp
@@ -78,7 +78,7 @@ void LocalPatchSource::update() {
 }
 
 LocalPatchSource::~LocalPatchSource() {
-    if ( !m_filename.isEmpty() ) {
+    if ( !m_command.isEmpty() && !m_filename.isEmpty() ) {
         QFile::remove( m_filename.toLocalFile() );
     }
 }
